+++ contrib/pylightning/lightning/lightning.py
@@ -329,31 +329,32 @@ class LightningRpc(UnixDomainSocketRpc):
     def close(self, peer_id, *args, **kwargs):
         """
         Close the channel with peer {id}, forcing a unilateral
-        close after {unilateraltimeout} seconds if non-zero.
+        close after {unilateraltimeout} seconds if non-zero, and
+        the to-local output will be sent to {destination}.
 
         Deprecated usage has {force} and {timeout} args.
         """
-        unilateraltimeout = None
 
         if 'force' in kwargs or 'timeout' in kwargs:
             return self._deprecated_close(peer_id, *args, **kwargs)
 
         # Single arg is ambigious.
-        if len(args) == 1:
+        if len(args) >= 1:
             if isinstance(args[0], bool):
                 return self._deprecated_close(peer_id, *args, **kwargs)
-            unilateraltimeout = args[0]
-        elif len(args) > 1:
-            return self._deprecated_close(peer_id, *args, **kwargs)
+            if len(args) == 2:
+                if args[0] is None and isinstance(args[1], int):
+                    return self._deprecated_close(peer_id, *args, **kwargs)
 
-        if 'unilateraltimeout' in kwargs:
-            unilateraltimeout = kwargs['unilateraltimeout']
+        def _close(peer_id, unilateraltimeout=None, destination=None):
+            payload = {
+                "id": peer_id,
+                "unilateraltimeout": unilateraltimeout,
+                "destination": destination
+            }
+            return self.call("close", payload)
 
-        payload = {
-            "id": peer_id,
-            "unilateraltimeout": unilateraltimeout
-        }
-        return self.call("close", payload)
+        return _close(peer_id, *args, **kwargs)
 
     def connect(self, peer_id, host=None, port=None):
         """
@@ -499,15 +500,11 @@ class LightningRpc(UnixDomainSocketRpc):
         }
         return self.call("feerates", payload)
 
-    def fundchannel(self, node_id, satoshi, feerate=None, announce=True, minconf=None, utxos=None):
-        """
-        Fund channel with {id} using {satoshi} satoshis with feerate
-        of {feerate} (uses default feerate if unset).
-        If {announce} is False, don't send channel announcements.
-        Only select outputs with {minconf} confirmations.
-        If {utxos} is specified (as a list of 'txid:vout' strings),
-        fund a channel from these specifics utxos.
-        """
+    def _deprecated_fundchannel(self, node_id, satoshi, feerate=None, announce=True, minconf=None, utxos=None):
+        warnings.warn("fundchannel: the 'satoshi' field is renamed 'amount' : expect removal"
+                      " in Mid-2020",
+                      DeprecationWarning)
+
         payload = {
             "id": node_id,
             "satoshi": satoshi,
@@ -518,16 +515,37 @@ class LightningRpc(UnixDomainSocketRpc):
         }
         return self.call("fundchannel", payload)
 
-    def fundchannel_start(self, node_id, satoshi, feerate=None, announce=True):
+    def fundchannel(self, node_id, *args, **kwargs):
         """
-        Start channel funding with {id} for {satoshi} satoshis
-        with feerate of {feerate} (uses default feerate if unset).
+        Fund channel with {id} using {amount} satoshis with feerate
+        of {feerate} (uses default feerate if unset).
         If {announce} is False, don't send channel announcements.
-        Returns a Bech32 {funding_address} for an external wallet
-        to create a funding transaction for. Requires a call to
-        'fundchannel_complete' to complete channel establishment
-        with peer.
+        Only select outputs with {minconf} confirmations.
+        If {utxos} is specified (as a list of 'txid:vout' strings),
+        fund a channel from these specifics utxos.
         """
+
+        if 'satoshi' in kwargs:
+            return self._deprecated_fundchannel(node_id, *args, **kwargs)
+
+        def _fundchannel(node_id, amount, feerate=None, announce=True, minconf=None, utxos=None):
+            payload = {
+                "id": node_id,
+                "amount": amount,
+                "feerate": feerate,
+                "announce": announce,
+                "minconf": minconf,
+                "utxos": utxos
+            }
+            return self.call("fundchannel", payload)
+
+        return _fundchannel(node_id, *args, **kwargs)
+
+    def _deprecated_fundchannel_start(self, node_id, satoshi, feerate=None, announce=True):
+        warnings.warn("fundchannel_start: the 'satoshi' field is renamed 'amount' : expect removal"
+                      " in Mid-2020",
+                      DeprecationWarning)
+
         payload = {
             "id": node_id,
             "satoshi": satoshi,
@@ -535,6 +553,31 @@ class LightningRpc(UnixDomainSocketRpc):
             "announce": announce,
         }
         return self.call("fundchannel_start", payload)
+
+    def fundchannel_start(self, node_id, *args, **kwargs):
+        """
+        Start channel funding with {id} for {amount} satoshis
+        with feerate of {feerate} (uses default feerate if unset).
+        If {announce} is False, don't send channel announcements.
+        Returns a Bech32 {funding_address} for an external wallet
+        to create a funding transaction for. Requires a call to
+        'fundchannel_complete' to complete channel establishment
+        with peer.
+        """
+
+        if 'satoshi' in kwargs:
+            return self._deprecated_fundchannel_start(node_id, *args, **kwargs)
+
+        def _fundchannel_start(node_id, amount, feerate=None, announce=True):
+            payload = {
+                "id": node_id,
+                "amount": amount,
+                "feerate": feerate,
+                "announce": announce
+            }
+            return self.call("fundchannel_start", payload)
+
+        return _fundchannel_start(node_id, *args, **kwargs)
 
     def fundchannel_cancel(self, node_id):
         """
@@ -891,7 +934,19 @@ class LightningRpc(UnixDomainSocketRpc):
 
         return self.call("withdraw", payload)
 
-    def txprepare(self, outputs, feerate=None, minconf=None, utxos=None):
+    def _deprecated_txprepare(self, destination, satoshi, feerate=None, minconf=None):
+        warnings.warn("txprepare now takes output arg: expect removal"
+                      " in Mid-2020",
+                      DeprecationWarning)
+        payload = {
+            "destination": destination,
+            "satoshi": satoshi,
+            "feerate": feerate,
+            "minconf": minconf,
+        }
+        return self.call("txprepare", payload)
+
+    def txprepare(self, *args, **kwargs):
         """
         Prepare a bitcoin transaction which sends to [outputs].
         The format of output is like [{address1: amount1},
@@ -901,13 +956,22 @@ class LightningRpc(UnixDomainSocketRpc):
         Outputs will be reserved until you call txdiscard or txsend, or
         lightningd restarts.
         """
-        payload = {
-            "outputs": outputs,
-            "feerate": feerate,
-            "minconf": minconf,
-            "utxos": utxos,
-        }
-        return self.call("txprepare", payload)
+        if 'destination' in kwargs or 'satoshi' in kwargs:
+            return self._deprecated_txprepare(*args, **kwargs)
+
+        if not isinstance(args[0], list):
+            return self._deprecated_txprepare(*args, **kwargs)
+
+        def _txprepare(outputs, feerate=None, minconf=None, utxos=None):
+            payload = {
+                "outputs": outputs,
+                "feerate": feerate,
+                "minconf": minconf,
+                "utxos": utxos,
+            }
+            return self.call("txprepare", payload)
+
+        return _txprepare(*args, **kwargs)
 
     def txdiscard(self, txid):
         """